src folder, LocalDateTime, Event and Planner unit tests
